+++ Cargo.toml
@@ -27,6 +27,7 @@ anyhow = "1"
 regex = "1"
 arborium = { version = "2", default-features = false, features = ["all-languages"] }
 ansi-to-tui = "8.0.1"
+open = "5.3.3"
 
 # The profile that 'dist' will build with
 [profile.dist]
