+++ src/diff/mod.rs
@@ -1,3 +1,4 @@
 pub mod expand;
+pub mod model;
 pub mod parser;
 pub mod renderer;
